feature: Реализация SessionParameter refactor: Вынес объекты метаданных по разным модулям
