+++ src/general.rs
@@ -1,9 +1,21 @@
+use std::path::Path;
+
 use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, PartialEq, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GeneralInfo {
+    #[serde(rename="xmlns:mdclass")]
+    pub md_class: String,
+    pub uuid: String,
+    pub name: StringValue,
+    pub comment: Option<StringValue>,
+}
 
 #[derive(Debug, Deserialize, PartialEq, Serialize)]
 pub struct StringValue {
     #[serde(rename="$value")]
-    pub value: Option<String>,
+    pub value: String,
 }
 
 #[derive(Debug, Deserialize, PartialEq, Serialize)]
@@ -16,4 +28,31 @@ pub struct Item {
 pub struct BooleanValue {
     #[serde(rename="$value")]
     body: bool,
+}
+
+#[derive(Debug, Deserialize, PartialEq, Serialize)]
+pub struct IntValue {
+    #[serde(rename="$value")]
+    body: i32,
+}
+
+pub trait MDObject {
+    fn new(root_path: &Path, name: String) -> Self;
+}
+
+pub fn read_objects<T: MDObject>(object_names: &Vec<StringValue>, root_path: &Path) -> Vec<T> {
+    object_names.iter()
+        .map(|full_name| &full_name.value)
+        .map(|name| extract_name_from_full_name(name))
+        .filter(|name| !name.is_empty())
+        .map(|name| T::new(root_path, name))
+        .collect()
+}
+
+fn extract_name_from_full_name(full_name: &String) -> String {
+    return if let Some(split) = full_name.split_once(".") {
+        split.1.to_string()
+    } else {
+        full_name.to_string()
+    };
 }
